split VaultList and ContextMenu into two separate controllers
